user_dao整理＆親クラス作って使い回せそうなメソッドを移植
